+++ src/wrapper/objects/jvalue.rs
@@ -1,9 +1,13 @@
+use std::char::{CharTryFromError, DecodeUtf16Error};
 use std::convert::{TryFrom, TryInto};
 use std::fmt::Debug;
 
 use log::trace;
 
 use crate::{errors::*, objects::JObject, signature::Primitive, sys::*};
+
+#[cfg(doc)]
+use crate::JNIEnv;
 
 /// A Java owned local reference or primitive value.
 ///
@@ -93,6 +97,22 @@ impl<'local> JValueOwned<'local> {
         self.borrow().c()
     }
 
+    /// Try to unwrap a Java `char` and then convert it to a Rust `char`.
+    ///
+    /// **Warning:** This conversion is likely to fail. Using it is not recommended. Prefer [`JValueGen::i_char`] where possible. See [`char_from_java`] for more information.
+    ///
+    /// # Errors
+    ///
+    /// This method can fail with two kinds of errors:
+    ///
+    /// * [`Error::WrongJValueType`]: `self` does not contain a Java `char`.
+    /// * [`Error::InvalidUtf16`]: `self` contains a Java `char`, but it is one half of a surrogate pair.
+    pub fn c_char(self) -> Result<char> {
+        let char = self.c()?;
+
+        char_from_java(char).map_err(|source| Error::InvalidUtf16 { source })
+    }
+
     /// Try to unwrap to a double.
     pub fn d(self) -> Result<jdouble> {
         self.borrow().d()
@@ -106,6 +126,20 @@ impl<'local> JValueOwned<'local> {
     /// Try to unwrap to an int.
     pub fn i(self) -> Result<jint> {
         self.borrow().i()
+    }
+
+    /// Try to unwrap a Rust `char` from a Java `int`. See [`char_from_java_int`] for details.
+    ///
+    /// # Errors
+    ///
+    /// This method can fail with two kinds of errors:
+    ///
+    /// * [`Error::WrongJValueType`]: `self` does not contain a Java `int`.
+    /// * [`Error::InvalidUtf32`]: `self` contains a Java `int`, but it is not a valid UTF-32 unit.
+    pub fn i_char(self) -> Result<char> {
+        let char = self.i()?;
+
+        char_from_java_int(char).map_err(|source| Error::InvalidUtf32 { char, source })
     }
 
     /// Try to unwrap to a long.
@@ -237,6 +271,22 @@ impl<'obj_ref> JValue<'obj_ref> {
         }
     }
 
+    /// Try to unwrap a Java `char` and then convert it to a Rust `char`.
+    ///
+    /// **Warning:** This conversion is likely to fail. Using it is not recommended. Prefer [`JValueGen::i_char`] where possible. See [`char_from_java`] for more information.
+    ///
+    /// # Errors
+    ///
+    /// This method can fail with two kinds of errors:
+    ///
+    /// * [`Error::WrongJValueType`]: `self` does not contain a Java `char`.
+    /// * [`Error::InvalidUtf16`]: `self` contains a Java `char`, but it is one half of a surrogate pair.
+    pub fn c_char(self) -> Result<char> {
+        let char = self.c()?;
+
+        char_from_java(char).map_err(|source| Error::InvalidUtf16 { source })
+    }
+
     /// Try to unwrap to a double.
     pub fn d(self) -> Result<jdouble> {
         match self {
@@ -261,6 +311,20 @@ impl<'obj_ref> JValue<'obj_ref> {
         }
     }
 
+    /// Try to unwrap a Rust `char` from a Java `int`. See [`char_from_java_int`] for details.
+    ///
+    /// # Errors
+    ///
+    /// This method can fail with two kinds of errors:
+    ///
+    /// * [`Error::WrongJValueType`]: `self` does not contain a Java `int`.
+    /// * [`Error::InvalidUtf32`]: `self` contains a Java `int`, but it is not a valid UTF-32 unit.
+    pub fn i_char(self) -> Result<char> {
+        let char = self.i()?;
+
+        char_from_java_int(char).map_err(|source| Error::InvalidUtf32 { char, source })
+    }
+
     /// Try to unwrap to a long.
     pub fn j(self) -> Result<jlong> {
         match self {
@@ -283,6 +347,11 @@ impl<'obj_ref> JValue<'obj_ref> {
             Self::Void => Ok(()),
             _ => Err(Error::WrongJValueType("void", self.type_name())),
         }
+    }
+
+    /// Converts a Rust `char` to a Java `int`. See [`char_to_java_int`] for details.
+    pub fn int_from_char(char: char) -> Self {
+        Self::Int(char_to_java_int(char))
     }
 }
 
@@ -383,6 +452,85 @@ impl<'obj_ref> TryFrom<JValue<'obj_ref>> for jchar {
     fn try_from(value: JValue) -> Result<Self> {
         value.c()
     }
+}
+
+/// Converts a Rust `char` to a Java `char`, if possible.
+///
+/// **Warning:** This conversion is likely to fail. Using it is not recommended. Prefer [`JValueGen::int_from_char`] where possible. See [`char_to_java`] for more information.
+impl<'local> TryFrom<char> for JValueOwned<'local> {
+    type Error = CharToJavaError;
+
+    fn try_from(value: char) -> std::result::Result<Self, Self::Error> {
+        Ok(Self::Char(char_to_java(value)?))
+    }
+}
+
+/// Converts a Rust `char` to a Java `char`, if possible.
+///
+/// **Warning:** This conversion is likely to fail. Using it is not recommended. Prefer [`JValueGen::int_from_char`] where possible. See [`char_to_java`] for more information.
+impl<'obj_ref> TryFrom<char> for JValue<'obj_ref> {
+    type Error = CharToJavaError;
+
+    fn try_from(value: char) -> std::result::Result<Self, Self::Error> {
+        Ok(Self::Char(char_to_java(value)?))
+    }
+}
+
+/// Converts a Java `char` to a Rust `char`, if possible. (Error-prone; see warning.)
+///
+/// **Warning:** Converting a single Java `char` to a Rust `char` is not recommended. This can only succeed for code points up to U+FFFF. Code points above that are represented in Java as *two* `char`s, each representing one half of a UTF-16 [surrogate pair], which this function cannot handle. If possible, use one of these alternatives instead:
+///
+/// * Use Java `int`s containing UTF-32. You can encode a Java `String` as a sequence of UTF-32 `int`s using its [`codePoints`] method, then use [`char_from_java_int`] to convert each one to a Rust `char`.
+///
+/// * Convert a Java `String` using [`JNIEnv::get_string`].
+///
+/// * Convert multiple Java `char`s at a time (such as in a Java `char[]` array) using [`char::decode_utf16`] (which this function is a simple wrapper around). That will properly convert any surrogate pairs among the Java `char`s.
+///
+/// # See Also
+///
+/// * [`JValueGen::c_char`], a wrapper around this function that unwraps [`JValueGen::Char`]
+/// * [`char_to_java`], the opposite of this function
+/// * [`char_from_java_int`], a UTF-32 alternative to this function that is unlikely to fail
+///
+/// # Errors
+///
+/// This function returns an error if the provided Java `char` is part of a UTF-16 surrogate pair, which cannot be converted to a Rust `char` by itself.
+///
+/// [`codePoints`]: https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/String.html#codePoints()
+/// [surrogate pair]: https://en.wikipedia.org/wiki/Surrogate_pair
+pub fn char_from_java(char: jchar) -> std::result::Result<char, DecodeUtf16Error> {
+    char::decode_utf16([char]).next().unwrap()
+}
+
+/// Converts a Rust `char` to a Java `char`, if possible. (Error-prone; see warning.)
+///
+/// **Warning:** Converting a Rust `char` to a single Java `char` is not recommended. This can only succeed for code points up to U+FFFF. Code points above that are represented in Java as *two* `char`s, each representing one half of a UTF-16 [surrogate pair], which this function cannot handle. If possible, use one of these alternatives instead:
+///
+/// * Use Java `int`s containing UTF-32. You can convert a Rust `char` to a Java `int` containing UTF-32 using [`char_to_java_int`], which never fails.
+///
+/// * Convert a Rust [`str`] to a Java `String` using [`JNIEnv::new_string`].
+///
+/// * Convert a Rust `char` to multiple Java `char`s at a time using [`char::encode_utf16`] (which this function is a wrapper around). That will properly generate surrogate pairs as needed.
+///
+/// # See Also
+///
+/// * [`JValueGen`]'s implementation of `TryFrom<char>`, a wrapper for this function that produces [`JValueGen::Char`]
+/// * [`char_from_java`], the opposite of this function
+/// * [`char_to_java_int`], a UTF-32 alternative to this function that never fails
+///
+/// # Errors
+///
+/// This function returns an error if the provided `char` cannot be represented in UTF-16 without a surrogate pair, and therefore cannot be converted to a single Java `char`.
+///
+/// [surrogate pair]: https://en.wikipedia.org/wiki/Surrogate_pair
+pub fn char_to_java(char: char) -> std::result::Result<jchar, CharToJavaError> {
+    if char.len_utf16() != 1 {
+        return Err(CharToJavaError { char });
+    }
+
+    let mut buf = [0u16; 1];
+    let buf: &mut [u16] = char.encode_utf16(&mut buf);
+    Ok(buf[0])
 }
 
 // jshort
@@ -499,6 +647,44 @@ impl<'obj_ref> TryFrom<JValue<'obj_ref>> for jint {
     fn try_from(value: JValue) -> Result<Self> {
         value.i()
     }
+}
+
+/// Converts a Rust `char` to a Java `int`.
+///
+/// This is the form expected or produced by certain Java APIs that process UTF-32 units, such as [`String.codePointAt`].
+///
+/// As discussed in [`char_to_java`], Rust `char` cannot always be converted to Java `char`, but can always be converted to Java `int`. This is the recommended way to pass a Rust `char` to Java code.
+///
+/// # See Also
+///
+/// * [`JValueGen::int_from_char`], a wrapper for this function that returns [`JValueGen::Int`]
+/// * [`char_from_java_int`], the opposite of this function
+/// * [`char_to_java`], an alternative to this function that converts to Java `char` but is likely to fail
+///
+/// [`String.codePointAt`]: https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/String.html#codePointAt(int)
+pub fn char_to_java_int(char: char) -> jint {
+    u32::from(char) as jint
+}
+
+/// Converts a Java `int` to a Rust `char`.
+///
+/// This is the form expected or produced by certain Java APIs that process UTF-32 units, such as [`String.codePointAt`].
+///
+/// As discussed in [`char_from_java`], Rust `char` cannot always be converted from Java `char`, but can always be converted from Java `int` (provided that the `int` contains a valid UTF-32 unit). This is the recommended way to receive a Rust `char` from Java code.
+///
+/// # See Also
+///
+/// * [`JValueGen::i_char`], a wrapper for this function that unwraps [`JValueGen::Int`]
+/// * [`char_to_java_int`], the opposite of this function
+/// * [`char_from_java`], an alternative to this function that converts from Java `char` but is likely to fail
+///
+/// # Errors
+///
+/// Returns an error if the Java `int` doesn't represent a valid UTF-32 unit.
+///
+/// [`String.codePointAt`]: https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/String.html#codePointAt(int)
+pub fn char_from_java_int(jint: jint) -> std::result::Result<char, CharTryFromError> {
+    char::try_from(jint as u32)
 }
 
 // jlong